# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bahar Premium Store API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface Product {
  id: number;
  name: string;
  nameAr?: string;
  description?: string;
  descriptionAr?: string;
  price: number;
  originalPrice?: number;
  imageUrl?: string;
  images?: string[];
  category?: string;
  stock: number;
  featured: boolean;
  origin?: string;
  createdAt?: string;
  updatedAt?: string;
}

export interface CreateProductInput {
  name: string;
  nameAr?: string;
  description?: string;
  descriptionAr?: string;
  price: number;
  originalPrice?: number;
  imageUrl?: string;
  images?: string[];
  category?: string;
  stock: number;
  featured?: boolean;
  origin?: string;
}

export interface UpdateProductInput {
  name?: string;
  nameAr?: string;
  description?: string;
  descriptionAr?: string;
  price?: number;
  originalPrice?: number;
  imageUrl?: string;
  images?: string[];
  category?: string;
  stock?: number;
  featured?: boolean;
  origin?: string;
}

export interface Offer {
  id: number;
  title: string;
  titleAr?: string;
  description?: string;
  descriptionAr?: string;
  productId?: number | null;
  product?: Product;
  products?: Product[];
  productIds?: number[];
  discountPercent?: number;
  offerPrice?: number;
  originalPrice?: number;
  imageUrl?: string;
  stockLimit?: number;
  stockRemaining?: number;
  expiresAt?: string;
  active: boolean;
  createdAt?: string;
}

export interface CreateOfferInput {
  title: string;
  titleAr?: string;
  description?: string;
  descriptionAr?: string;
  productId?: number | null;
  productIds?: number[];
  discountPercent?: number;
  offerPrice?: number;
  originalPrice?: number;
  imageUrl?: string;
  stockLimit?: number;
  stockRemaining?: number;
  expiresAt?: string;
  active: boolean;
}

export interface UpdateOfferInput {
  title?: string;
  titleAr?: string;
  description?: string;
  descriptionAr?: string;
  productId?: number | null;
  productIds?: number[];
  discountPercent?: number;
  offerPrice?: number;
  originalPrice?: number;
  imageUrl?: string;
  stockLimit?: number;
  stockRemaining?: number;
  expiresAt?: string;
  active?: boolean;
}

export interface OrderItem {
  productId: number;
  productName: string;
  quantity: number;
  unitPrice: number;
  totalPrice: number;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  processing: "processing",
  shipped: "shipped",
  delivered: "delivered",
  canceled: "canceled",
} as const;

export interface Order {
  id: number;
  customerName: string;
  customerPhone: string;
  customerAddress: string;
  customerEmail?: string;
  items: OrderItem[];
  totalPrice: number;
  status: OrderStatus;
  notes?: string;
  createdAt?: string;
  updatedAt?: string;
}

export interface CreateOrderInput {
  customerName: string;
  customerPhone: string;
  customerAddress: string;
  customerEmail?: string;
  items: OrderItem[];
  totalPrice: number;
  notes?: string;
}

export type UpdateOrderInputStatus =
  (typeof UpdateOrderInputStatus)[keyof typeof UpdateOrderInputStatus];

export const UpdateOrderInputStatus = {
  pending: "pending",
  processing: "processing",
  shipped: "shipped",
  delivered: "delivered",
  canceled: "canceled",
} as const;

export interface UpdateOrderInput {
  status?: UpdateOrderInputStatus;
  notes?: string;
}

export interface Customer {
  id: number;
  name: string;
  phone: string;
  email?: string;
  address?: string;
  totalOrders?: number;
  totalSpent?: number;
  createdAt?: string;
}

export interface UpdateCustomerInput {
  name?: string;
  phone?: string;
  email?: string;
  address?: string;
}

export interface SiteStatus {
  isOpen: boolean;
  closedMessage?: string;
  reopensAt?: string | null;
}

export interface AdminLoginInput {
  password: string;
}

export interface AdminLoginResponse {
  success: boolean;
  message?: string;
}

export interface AdminInfo {
  authenticated: boolean;
  role?: string;
}

export interface UploadResponse {
  url: string;
  filename?: string;
}

export interface MultiUploadResponse {
  urls: string[];
  filenames?: string[];
}

export interface DashboardStats {
  totalOrders: number;
  totalRevenue: number;
  totalProducts: number;
  totalCustomers: number;
  pendingOrders: number;
  recentOrders: Order[];
}

export interface Message {
  id: number;
  name: string;
  email: string;
  subject?: string | null;
  message: string;
  read: boolean;
  createdAt?: string | null;
}

export interface CreateMessageInput {
  name: string;
  email: string;
  subject?: string;
  message: string;
}

export type GetProductsParams = {
  category?: string;
  featured?: boolean;
};

export type UploadImageBody = {
  image?: Blob;
};

export type UploadImagesBody = {
  images?: Blob[];
};