# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Bahar Premium Store API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: getProducts
      tags: [products]
      summary: Get all products
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: featured
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductInput"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update a product (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductInput"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /admin/site-status:
    get:
      operationId: getSiteStatus
      tags: [admin]
      summary: Get site open/closed status
      responses:
        "200":
          description: Site status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteStatus"
    put:
      operationId: updateSiteStatus
      tags: [admin]
      summary: Update site status (open/close with optional timer)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SiteStatus"
      responses:
        "200":
          description: Updated site status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteStatus"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        nameAr:
          type: string
        description:
          type: string
        descriptionAr:
          type: string
        price:
          type: number
        originalPrice:
          type: number
        imageUrl:
          type: string
        images:
          type: array
          items:
            type: string
        category:
          type: string
        stock:
          type: integer
        featured:
          type: boolean
        origin:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - price
        - stock
        - featured
    CreateProductInput:
      type: object
      properties:
        name:
          type: string
        nameAr:
          type: string
        description:
          type: string
        descriptionAr:
          type: string
        price:
          type: number
        originalPrice:
          type: number
        imageUrl:
          type: string
        images:
          type: array
          items:
            type: string
        category:
          type: string
        stock:
          type: integer
        featured:
          type: boolean
        origin:
          type: string
      required:
        - name
        - price
        - stock
    UpdateProductInput:
      type: object
      properties:
        name:
          type: string
        nameAr:
          type: string
        description:
          type: string
        descriptionAr:
          type: string
        price:
          type: number
        originalPrice:
          type: number
        imageUrl:
          type: string
        images:
          type: array
          items:
            type: string
        category:
          type: string
        stock:
          type: integer
        featured:
          type: boolean
        origin:
          type: string
...
    SiteStatus:
      type: object
      required: [isOpen]
      properties:
        isOpen:
          type: boolean
        closedMessage:
          type: string
        reopensAt:
          type: string
          nullable: true
...