# lib/api-zod/src/generated/types/siteStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bahar Premium Store API
 * OpenAPI spec version: 0.1.0
 */

export interface SiteStatus {
  isOpen: boolean;
  closedMessage?: string;
  reopensAt?: string | null;
}